discogs_credits: extract MB relationship-editor functions
Issue #32 — the three MB-React-shaped functions move out:

  src/editor-state.js
    waitForMBEditor       poll until MB.relationshipEditor mounts
    dispatchRelationship  feed one rel into MB's React reducer
                          (entity0/1 swap, attr-tree iteration for the
                          log line, credit routing)
    buildAttributes       turn the script's mixed attribute representation
                          into MB's internal ImmutableTree via
                          tree.fromDistinctAscArray; resolves names against
                          MB.linkedEntities.link_attribute_type, drops
                          unknowns with a WARN (issue #3 guard)

~150 lines moved. Deps already extracted (pageWindow, REL_TEMPLATE,
addLogLine). No behavior change; full 12-fixture suite green.
